CodeClean_sortByPriceRange
